Burgers' equation implementation
It's a very simple and detailed keras implementation of PINN
solving Burgers' equation.

There are many spikes in the loss plot. It seems training may not be
stable. Raissi never show their loss plots. I'm able to reproduce their
results, but these spikes need to be investigated. Because you may end
up with bad results if you stop training in one of these spikes. Or not?